# Fix tests + optimize SourceLine()