- Fix #21, fix #22 
- Optimize `(frame *StackFrame) SourceLine()`

# Benchmarks:

**Before optimization**

```
➜  errors git:(master) ✗ go test -bench . ./...          
goos: darwin
goarch: amd64
pkg: github.com/Thiht/errors
BenchmarkStackFormat-4              2511            417541 ns/op          391870 B/op        207 allocs/op
PASS
ok      github.com/Thiht/errors 1.107s
```

**After optimization**

```
➜  errors git:(master) ✗ go test -bench . ./...
goos: darwin
goarch: amd64
pkg: github.com/Thiht/errors
BenchmarkStackFormat-4              3891            299648 ns/op           53850 B/op        187 allocs/op
PASS
ok      github.com/Thiht/errors 2.173s
```

20 less allocations per operation
~8 times less memory usage per operation