@ConradIrwin 2020-04-13
Thanks for this! :gem:

FWIW, we've had problems with the tests breaking every go version; but if they work for now, happy to merge this.